+++ router/h2/src/e2e/scala/io/buoyant/router/h2/FlowControlEndToEndTest.scala
@@ -11,36 +11,37 @@ import scala.collection.mutable.ListBuffer
 class FlowControlEndToEndTest
   extends FunSuite
   with ClientServerHelpers {
-  setLogLevel(Level.OFF)
 
   test("client/server request flow control") {
     val streamP = new Promise[Stream]
     val server = { req: Request =>
-      streamP.setValue(req.data)
-      Response(Status.Ok, Stream.Nil)
+      streamP.setValue(req.stream)
+      Response(Status.Ok, Stream.empty())
     }
     withClient(server) { client =>
       val writer = Stream()
       val req = Request("http", Method.Get, "host", "/path", writer)
       val rsp = await(client(req))
       assert(rsp.status == Status.Ok)
-      testFlowControl(reader(await(streamP)), writer)
+      testFlowControl(await(streamP), writer)
     }
   }
 
   test("client/server response flow control") {
-    val writer = Stream()
-    withClient(_ => Response(Status.Ok, writer)) { client =>
-      val req = Request("http", Method.Get, "host", "/path", Stream.Nil)
-      val rsp = await(client(req))
-      assert(rsp.status == Status.Ok)
-      testFlowControl(reader(rsp.data), writer)
-    }
+    try {
+      val writer = Stream()
+      withClient(_ => Response(Status.Ok, writer)) { client =>
+        val req = Request("http", Method.Get, "host", "/path", Stream.empty())
+        val rsp = await(client(req))
+        assert(rsp.status == Status.Ok)
+        testFlowControl(rsp.stream, writer)
+      }
+    } finally setLogLevel(Level.OFF)
   }
 
   val WindowSize = 65535
 
-  def testFlowControl(reader: Stream.Reader, writer: Stream.Writer) = {
+  def testFlowControl(reader: Stream, writer: Stream.Writer) = {
     // The first frame is too large to fit in a window. It should not
     // be released until all of the data has been flushed.  This
     // cannot happen until the reader reads and releases some of the
@@ -51,8 +52,8 @@ class FlowControlEndToEndTest
 
     log.debug("offering 2 frames with %d", 2 * WindowSize)
     val wrote0 = writer.write(frame0)
-    val wrote1 = writer.write(frame1)
-    val wrote2 = writer.write(frame2)
+    val wrote1 = wrote0.before(writer.write(frame1))
+    val wrote2 = wrote1.before(writer.write(frame2))
     assert(!wrote0.isDefined && !wrote1.isDefined && !wrote2.isDefined)
 
     // Read a full window, without releasing anything.
@@ -61,12 +62,13 @@ class FlowControlEndToEndTest
     while (read < WindowSize) {
       log.debug("~~~ reading %d/%d", read, WindowSize)
       await(reader.read()) match {
-        case _: Frame.Trailers => fail("unexpected trailers")
         case d: Frame.Data =>
           // frames += d
           read += d.buf.length
           log.debug("~~~ read %d = %d/%d", d.buf.length, read, WindowSize)
           await(d.release())
+
+        case f => fail(s"unexpected frame: $f")
       }
     }
     assert(read == WindowSize)
@@ -79,11 +81,12 @@ class FlowControlEndToEndTest
         read, 2 * WindowSize, wrote1.isDefined)
       if (read > WindowSize + 1024) assert(wrote0.isDefined)
       await(reader.read()) match {
-        case _: Frame.Trailers => fail("unexpected trailers")
         case d: Frame.Data =>
           read += d.buf.length
           log.debug("reader releasing %dB from 1 frame", d.buf.length)
           await(d.release())
+
+        case f => fail(s"unexpected frame: $f")
       }
     }
     assert(read == 2 * WindowSize)
@@ -94,12 +97,13 @@ class FlowControlEndToEndTest
       log.debug("reader reading more after %dB/%dB, released=%s",
         read, 3 * WindowSize, wrote2.isDefined)
       await(reader.read()) match {
-        case _: Frame.Trailers => fail("unexpected trailers")
         case d: Frame.Data =>
           read += d.buf.length
           log.debug("reader releasing %dB from 1 frame", d.buf.length)
           await(d.release())
           assert(d.isEnd == (read == 3 * WindowSize))
+
+        case f => fail(s"unexpected frame: $f")
       }
     }
     assert(read == 3 * WindowSize)