+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 * Change default value of `attemptTTwitterUpgrade` to `false`
 * The `io.l5d.consul` and `io.l5d.k8s` namers are no longer experimental
+* Fix the `h2` protocol to gracefully handle connection loss and
+  stream interruption.
 
 ## 0.8.3